server/requirements.txt
dj-rest-auth
django-allauth
django
djangorestframework
pytz
djangorestframework-simplejwt
django-cors-headers
drf-yasg
pillow
elasticsearch>=2.0.0,<3.0.0
drf-haystack
django-debug-toolbar